wip: typing, renaming files to fix package import
