+++ src/breds/pattern.py
@@ -5,7 +5,10 @@ import uuid
 from typing import Set, Any, Optional
 
 from breds.config import Config
-from breds.bredstuple import BREDSTuple
+from breds.breds_tuple import BREDSTuple
+
+# from config import Config
+# from breds_tuple import BREDSTuple
 
 
 class Pattern:  # pylint: disable=too-many-instance-attributes
